backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
python-dotenv
python-docx
PyMuPDF
aiofiles
gTTS
TTS
